Differential analysis: Add ignored file difference properties to DFXMLObject; add NIST disclaimer
This fixes some confusing behavior.  For example, take this call:

    python3 make_differential_dfxml.py --debug --ignore atime --ignore data_brs ../samples/difference_test_0.xml ../samples/difference_test_1.xml

The file `i_will_be_accessed.txt` is excluded from the difference report
with both of those `--ignore` flags supplied.  However, if only the
`atime` ignore flag is supplied, the file would still be reported in the
output fileobject difference stream, as a (metadata-)changed file
(because the byte runs are changed), but with the atime field flagged as
changed, even though it was requested to be excluded from difference
reporting.

This commit moves the file difference set to the DFXMLObject, to create
a record of properties that were subtracted from the difference
reporting.

The same modification should probably be done with the property mask set
as well.

Reported-by: Dianne Dietrich <dd388@cornell.edu>
Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>