Objects.py: Add LibraryObject reading and writing, and unit tests
This update adds libraries to capture basic version information on
runtime and build-time libraries.  The new unit tests show this can be a
way to check for libraries even if only a name match is desired, and can
also be a way to record the Python major version (esp. 2 vs. 3) used
when generating a DFXML document.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>