Revise appearance of "modified" differential annotation
This patch adds SHA256 to the list of properties that, on differing
between two DFXML manifests, trigger a file being considered to be
"modified".  This pach also removes "mtime" from the list of triggering
fields.

Rationale on mtime:  mtime may denote a modification having taken place
and then being reverted later back to original contents; or, a timestamp
update as with the `touch` utility.  However, if content hashes match
before and after such an operation, then a file is not modified
according to the states being compared.  This kind of ultimately-nop
modification may be better investigated by, for example, an application
scanning for deleted editor files on detecting an mtime change.

As with all changes, feedback is welcome on whether mtime changing
should trigger a "modified" declaration.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>