walk_to_dfxml.py: Add property-ignoring flag
This builds on @kieranjol's suggestion to implement an ignore-hashes
flag, discussed in this Pull Request:
https://github.com/simsong/dfxml/pull/28

A more generic property-ignoring flag is implemented here to allow
ignoring any property of a `FileObject`.  A further selector, the `@`
character, can ignore the property only for certain name types.  The
unit test `verify_walk_ignore_genprops.py` demonstrates how to ignore
the `mtime` of a directory.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>