objects.py: Add error property to disk image object
pop_poststream_elements is similar to the other copies of the function,
except for external-namespace element handling.  The DFXML schema
doesn't have a provision for external-namespace elements at the end of
the child element list at the moment.  If determined to be a bug in the
schema, that can be added in a patch-level release when found necessary.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>