objects.py tests: Reduce set of tests run by Makefile and pytest
When Simson transitioned the test framework to be primarily based on
pytest, some of the tests I'd written to work via `make check` were left
in place.  Now that there is a better function for file I/O testing,
probably all of these tests can be moved to run via pytest instead of
`make check`.  This commit picks off some of the low-hanging fruit of
this machine work reduction.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>