setup.py: Fix deployment of modules with pip (#33)
* setup.py: Add unit test for setup.py pip deployment

This test was introduced to confirm the present state of `pip` and
`setup.py` being nonfunctional.

* setup.py: Use setuptools instead of distutils

Per the Python packaging recommendations:
https://packaging.python.org/guides/tool-recommendations/

This also exposes the function `find_packages`, which appears to be
necessary to export the DFXML modules.  Per:
https://packaging.python.org/tutorials/packaging-projects/#packaging-your-project

This patch is necessary, but insufficient, to fix the unit test for
`pip` and `setup.py`.

* setup.py: Use setuptools.find_packages

This patch fixes the unit test for `pip` and `setup.py`, confirming
export of `dfxml.py`, `objects.py`, and the other `.py` files under
`./dfxml`.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>

* setup.py: Remove py_modules

The `py_modules` line, an artifact of `distutils`, is not necessary
to enable `pip install .../dfxml/python`.  It was not working in
this file at this point in the source directory hierarchy.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>