objects.py: Shore up storage layer transitions
Some storage layer transitions were found unimplemented when testing an
objects.py-based tool.  This patch adds the missing transitions that
appear to complete support of the layers added in DFXML Schema 1.2.0.
A test revision coming in the next patch shows just how these appear to
complete that support.

There are some TODOs put in that note some assumed transition contexts.
It would be more correct to add parser states that capture when the
element stream is in a "Child-list" mode, tied to the container class;
but that increases the parser code complexity significantly.  As a
matter of pragmatism, those TODOs will remain there until a sufficiently
complex sample comes along.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>